Improve entity analysis and set automatic threshold for early stopping
1. Use built-in tokenizer to first split the input text into different
parts, and then feed it into the analyzer. 

2. The person/location model can only recognize records that the first
char is upper case. Added input preprocessing for these two models. 

3. add convergence test. In current implementation, the convergence test
check if the current estimated type with maximum probability is still
the maximum after another read of data. If the answer is yes, then it
skip the entity analyzing part. Otherwise, it will further feed the
records to entity analyzer until the above stopping condition meets.